chore: refactor go types (#1925)
This change extracts GoProperty and GoStruct from the Struct and GoClass code, as they had shared logic.


---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0
