+++ packages/jsii-pacmak/lib/targets/golang/types/type-field.ts
@@ -1,10 +1,10 @@
-import { GoClass, Interface, GoTypeRef } from './index';
+import { GoClass, Interface, Struct, GoTypeRef } from './index';
 
 /*
  * Structure for Class and Interface methods. Useful for sharing logic for dependency resolution
  */
 export interface TypeField {
   name: string;
-  parent: GoClass | Interface;
+  parent: GoClass | Interface | Struct;
   references?: GoTypeRef;
 }
